Tidied up and changed the structure of proof.
More restructuring is still required; things left in mod.rs need to be
moved.
Higher level documentation should also be added.

Another todo is adding a proof trait that other proof concepts can
implement.
