+++ src/main.rs
@@ -13,7 +13,7 @@ use verify::verify;
 
 fn main() {
     //for _ in 0..100000 {
-        proof::examples();
+    proof::examples::examples();
     //}
     return;
